Move Tracker protocol to its own file
